Fix stac_requset type
